chore: rename folder, files, component
