+++ frontend/src/components/login/login-input/index.ts
@@ -1,0 +1,1 @@
+export { default } from './LoginInput'
