--- frontend/src/components/sign-in/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SignIn'
